Add files via upload
Modularization of the menu bar;
Changing the "type" variable to "optionDriectory";
Comments;
Modifications for compare.setOnAction();
