+++ DifferenceTool/src/gui/DiffGUI.java
@@ -1,22 +1,7 @@
 package gui;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.HashMap;
-
 import javafx.application.Application;
-import javafx.application.Platform;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.scene.Scene;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.Menu;
-import javafx.scene.control.MenuBar;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.RadioMenuItem;
-import javafx.scene.control.ToggleGroup;
-import javafx.scene.control.TreeView;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
 import javafx.stage.Stage;
@@ -25,153 +10,36 @@ public class DiffGUI extends Application {
 
 	Stage window;
 	Scene primaryScene;
-	GridPane directoryPane = new GridPane();
 	GridPane filePane = new GridPane();
-	DirectoryViewer directoryView;
-
-	boolean type = false; // For file false; For directories true
+	BorderPane layout = new BorderPane();
+	
 
 	public void start(Stage primaryStage) throws Exception {
+		
 		window = primaryStage;
 		window.setMaximized(true);
 
-		BorderPane layout = new BorderPane();
 		primaryScene = new Scene(layout);
-
-		MenuBar menuBar = new MenuBar();
-		// --- Menu File
-		Menu menuFile = new Menu("File");
-		Menu menuOptions = new Menu("Options");
-		Menu menuHelp = new Menu("Help");
-		MenuItem about = new MenuItem("About");
-		menuHelp.getItems().add(about);
-
-		menuBar.getMenus().addAll(menuFile, menuOptions, menuHelp);
-		MenuItem compare = new MenuItem("Compare");
-		menuFile.getItems().add(compare);
-		RadioMenuItem file = new RadioMenuItem("File");
-		RadioMenuItem folder = new RadioMenuItem("Directory");
-		ToggleGroup tGroup = new ToggleGroup();
-
-		layout.setTop(menuBar);
-
-		file.setToggleGroup(tGroup);
-		folder.setToggleGroup(tGroup);
-
-		file.setOnAction(new EventHandler<ActionEvent>() {
-
-			@Override
-			public void handle(ActionEvent event) {
-
-				layout.getChildren().remove(directoryPane);
-				type = false;
-
-				FileViewer fileView = FileViewer.getInstance();
-				filePane = fileView.getDefaultFileView(window);
-
-				layout.setCenter(filePane);
-
-			}
-
-		});
-
-		folder.setOnAction(new EventHandler<ActionEvent>() {
-
-			@Override
-			public void handle(ActionEvent event) {
-
-				layout.getChildren().remove(filePane);
-				type = true;
-
-				directoryView = DirectoryViewer.getInstance();
-				directoryPane = directoryView.getDirectoryView(window);
-
-				layout.setCenter(directoryPane);
-
-			}
-
-		});
-
-		// By default, the tool is set up for files type
-		file.setSelected(true);
-
-		menuOptions.getItems().addAll(file, folder);
-
+		
+		/**
+		 * Adding the menu bar on the top of layout
+		 */
+		MenuBarFactory menuBar = new MenuBarFactory();
+		layout.setTop(menuBar.getMenuBar(layout, window));
+		
+		/**
+		 * As default, the files comparison pane is set
+		 */
 		FileViewer fileView = FileViewer.getInstance();
 		filePane = fileView.getDefaultFileView(window);
-
+		
 		layout.setCenter(filePane);
 
 		window.setTitle("Diff Tool");
 		window.setScene(primaryScene);
 		window.show();
 
-		about.setOnAction(new EventHandler<ActionEvent>() {
-
-			public void handle(ActionEvent event) {
-				Alert alert = new Alert(AlertType.INFORMATION);
-				alert.setTitle("About Diff Tool");
-				alert.setHeaderText("DiffTool is the visual file comparison (diff). Use it to compare source code,"
-						+ " \n web pages and other text files with native application performance.");
-				alert.setContentText("From \"Options\" menu item you can select what type of comparison do you need."
-						+ "\nFor file/folder selection, you should paste the absolute path and that click on "
-						+ "\n\"Get path\" button for importing the file or use just the button to select "
-						+ "\nmanually.");
-
-				alert.showAndWait();
-
-			}
-
-		});
-
-		compare.setOnAction(new EventHandler<ActionEvent>() {
-
-			public void handle(ActionEvent event) {
-				DirectoryViewer dw = DirectoryViewer.getInstance();
-				if (type && ((dw.getFirstDir().isDirectory() && (dw.getLastDir().isDirectory())))) {
-					TreeView<Object> tv1 = directoryView.getLeftHelper().getTreeView();
-					TreeView<Object> tv2 = directoryView.getRightHelper().getTreeView();
-					DirectoriesComaprison compare = new DirectoriesComaprison(dw.getFirstDir(), dw.getLastDir(), tv1,
-							tv2);
-					try {
-						Platform.runLater(new Runnable() {
-							@Override
-							public void run() {
-								try {
-									compare.getDiff(compare.getLeftDir(), compare.getRightDir());
-								} catch (IOException e) {
-									e.printStackTrace();
-								}
-								
-								HashMap<String, String> newMap = compare.getMapDiff();
-
-								DirectoriesDifferences dd = new DirectoriesDifferences();
-								TreeViewHelper treeHelper = new TreeViewHelper();
-								TreeView<Object> auxView = directoryView.rightHelper.getTreeView();
-								
-								treeHelper.setTreeView((dd.markDifferences(auxView, newMap,
-										new File(directoryView.getRightText().toString()))));
-								dw.setRightHelper(treeHelper);
-
-								auxView = directoryView.leftHelper.getTreeView();
-								treeHelper.setTreeView(dd.markDifferences(auxView, newMap,
-										new File(directoryView.getLeftText().toString())));
-								dw.setLeftHelper(treeHelper);
-							}
-						});
-					} catch (Exception e) {
-						e.printStackTrace();
-					}
-
-				} else if (!type) {
-					filePane = FileViewer.getInstance().changePaneToDiffs(window);
-					layout.setCenter(filePane);
-				}
-
-			}
-
-		});
-
+		
 	}
 
 	public static void main(String[] args) {
